perf: find 성능 개선
+ NPE 문제 생길 수 있는 문제 개선
+ 사용하지 않는 import 삭제

+++ HRBank/src/main/java/org/yebigun/hrbank/domain/binaryContent/storage/BinaryContentStorageImpl.java
@@ -19,6 +19,8 @@ import java.io.InputStream;
 import java.nio.file.*;
 import java.util.List;
 import java.util.NoSuchElementException;
+import java.util.Set;
+import java.util.stream.Collectors;
 import java.util.stream.Stream;
 
 /**
@@ -64,7 +66,7 @@ public class BinaryContentStorageImpl implements BinaryContentStorage {
     public int deleteUnusedFiles() {
         try (DirectoryStream<Path> stream = Files.newDirectoryStream(root)) {
             // 1. DB에서 사용 중인 파일 이름 목록 수집
-            List<String> validFileNames = binaryContentRepository.findAll()
+            Set<String> validFileNames = binaryContentRepository.findAll()
                 .stream()
                 .flatMap(b -> {
                     String ext = getExtension(b.getFileName());
@@ -74,15 +76,17 @@ public class BinaryContentStorageImpl implements BinaryContentStorage {
                         return Stream.of(b.getId() + ext);
                     }
                 })
-                .toList();
+                .collect(Collectors.toSet());
 
             int deletedCount = 0;
             for (Path file : stream) {
-                String fileName = file.getFileName().toString();
-                if (!validFileNames.contains(fileName)) {
-                    Files.deleteIfExists(file);
-                    deletedCount++;
-                    log.info("삭제된 미사용 파일: {}", fileName);
+                if (Files.isRegularFile(file)) {
+                    String fileName = file.getFileName().toString();
+                    if (!validFileNames.contains(fileName)) {
+                        Files.deleteIfExists(file);
+                        deletedCount++;
+                        log.info("삭제된 미사용 파일: {}", fileName);
+                    }
                 }
             }
             return deletedCount;
